requirements.txt
PyYAML ~= 3.0
